+++ hmailserver/source/Server/SMTP/SMTPConnection.h
@@ -197,7 +197,7 @@ namespace HM
       
       DWORD message_start_tc_;
 
-      int max_message_size_kb_;
+      size_t max_message_size_kb_;
       // Maximum message size in KB.
 
       String helo_host_;
